requirements.txt
# Core
flask==3.0.0
python-dotenv==1.0.0

# WhatsApp/Twilio
twilio==8.11.0

# Google Calendar
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# Database
pymongo==4.6.0
redis==5.0.1

# Time handling
pytz==2023.3

# Development
pytest==7.4.3
pytest-cov==4.1.0

# Production server
gunicorn==21.2.0

# Utilities
requests==2.31.0